==> fpro_code/1_gen_sessions_fpro.ipynb <==
import numpy as np
import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
import pyarrow
from pandas import Timestamp
import itertools
from datetime import datetime
import datetime
import time
from tqdm import tqdm
from datetime import datetime
from joblib import Parallel, delayed
import collections
import random
import pyarrow.parquet as pq

# my credential file
#key_path = "/Users/yuxuanyang/Downloads/encoded-adviser-131008-f2e09e85c3b3.json"
key_path = "/Users/yuxuanyang/Downloads/encoded-adviser-131008-7129911936a3.json"

credentials = service_account.Credentials.from_service_account_file(
    key_path, scopes=["https://www.googleapis.com/auth/cloud-platform"],
)

client = bigquery.Client(credentials=credentials, project=credentials.project_id)

query= """
    select *
    from `encoded-adviser-131008.yang_20230130.20000users`
"""
query_job = client.query(query)  # Make an API request.
users = query_job.to_dataframe()
pd.to_pickle(users, '/Users/yuxuanyang/Downloads/DSIN-master/fpro_data/sample_users.pkl')

query= """
    select *
    from `encoded-adviser-131008.yang_20230130.apps`
"""
query_job = client.query(query)  # Make an API request.
apps = query_job.to_dataframe()
pd.to_pickle(apps, '/Users/yuxuanyang/Downloads/DSIN-master/fpro_data/apps.pkl')

query= """
    select a.user_id as user_id, a.datetime as datetime, a.app_id as app_id, b.cate as cate, b.store as store
    from `encoded-adviser-131008.yang_20230130.logs_8days_20000users` a
    left join `encoded-adviser-131008.yang_20230130.apps` b
    on a.app_id = b.app_id
    order by cate
"""
query_job = client.query(query)  # Make an API request.
user_logs = query_job.to_dataframe()
#pd.to_pickle(user_logs, '/Users/yuxuanyang/Downloads/DSIN-master/fpro_data/sample_user_logs.pkl')

query= """
    select a.user_id as user_id, a.app_id as app_id, count(a.datetime) as access_count
    from `encoded-adviser-131008.yang_20230130.foreground_log_201906` a
    inner join `encoded-adviser-131008.yang_20230130.20000users` b
    on a.user_id = b.user_id
    group by user_id, app_id
    order by access_count
"""
query_job = client.query(query)  # Make an API request.
user_apps = query_job.to_dataframe()
#pd.to_pickle(user_logs, '/Users/yuxuanyang/Downloads/DSIN-master/fpro_data/sample_user_logs.pkl')

def unix_to_timestamp(ut):
    return datetime.fromtimestamp(ut).strftime('%Y-%m-%d %H:%M:%S')
def timestamp_to_unix(ts):
    return time.mktime(ts.timetuple())

user_logs = pd.read_parquet('/Volumes/PortYoox/Deep_learning_models/DSIN-master/fpro_data/sample_user_logs.parquet')

user_logs = user_logs.reset_index().drop(['index','cate', 'store'],axis = 1)
user_logs 

def gen_negative_samples(user, log, sample_by_popular):
    log = log.reset_index()
    # for each access log, the negative samples were generated by the other apps in the user's device but did not click
    app_counter = collections.Counter(log['app_id'].tolist())
    app_id_list = []
    time_list = []
    for i in range(len(log)):
        curr_app = log.iloc[i]['app_id']
        time_stamp = log.iloc[i]['time_stamp']
        if sample_by_popular == False:
            other_apps = [a for a in list(app_counter.keys()) if a != curr_app]
            if other_apps:
                sample_negative_app = random.choice(other_apps)
                app_id_list.append(sample_negative_app)
                time_list.append(time_stamp)
              #  negative_df.loc[len(negative_df)]= [user, sample_negative_app, time_stamp, 0]
        elif sample_by_popular == True:
            other_apps = [a for a in list(app_counter.keys()) if a != curr_app]
            popular_sorted_apps = sorted(other_apps, key = lambda x: app_counter[x])
            if popular_sorted_apps:
                sample_negative_app = popular_sorted_apps[-1]
                app_id_list.append(sample_negative_app)
                time_list.append(time_stamp)
            #   negative_df.loc[len(negative_df)]= [user, sample_negative_app, time_stamp, 0]
    l = len(app_id_list)
    negative_df = pd.DataFrame({'user_id': [user]*l,
                                'app_id': app_id_list,
                                'time_stamp': time_list,
                                'clk': [0]*l})
    return negative_df

def applyParallel(dfGrouped, func, sample_by_popular):
    res = Parallel(n_jobs= -1)(delayed(func)(name, group, sample_by_popular) \
                                 for name, group in tqdm(dfGrouped))
    return pd.concat(res)


negative_samples_by_random = applyParallel(user_logs.groupby('user_id'), gen_negative_samples, False)

negative_samples_by_random = negative_samples_by_random.reset_index().drop(['index'],axis = 1)

pd.to_pickle(negative_samples_by_random, '../fpro_data/negative_samples_by_random.pkl')

negative_samples_by_popular = applyParallel(user_logs.groupby('user_id'), gen_negative_samples, True)

negative_samples_by_popular = negative_samples_by_popular.reset_index().drop(['index'],axis = 1)
negative_samples_by_popular 

pd.to_pickle(negative_samples_by_popular , '../fpro_data/negative_samples_by_popular.pkl')

user_logs.replace(to_replace=[None], value= ['nan'], inplace=True)

import numpy as np
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm
from sklearn.ensemble import IsolationForest
from pandas.testing import assert_frame_equal

FRAC = 0.25
user_logs = pd.read_pickle('../fpro_data/sample_user_logs.pkl')
user_logs.replace(to_replace=[None], value= ['nan'], inplace=True)
def gen_intervals(user, user_log, groupby_sdt):
    sorted_log = user_log.sort_values(by = ['time_stamp'])
    if len(sorted_log) > 1:
        intervals = {}
        time1 = sorted_log.time_stamp.to_numpy()[:-1]
        time2= sorted_log.time_stamp.to_numpy()[1:]
        name = sorted_log[groupby_sdt].to_numpy()[:-1]
        diffs = time2 - time1 
        return pd.DataFrame({groupby_sdt: name, 'interval': diffs})

def applyParallel(dfGrouped, func, groupby_sdt):
    res = Parallel(n_jobs= -1)(delayed(func)(name, group, groupby_sdt) for name, group in tqdm(dfGrouped))
#log.iloc[:100].groupby('user').apply(gen_intervals)
    return pd.concat(res)
#删除异常值 用异常分数的0.5作为阀值
def iforest_remove_outliner(name, intervals_groupby, groupby_sdt):
    sll = intervals_groupby.interval.tolist()

    X = np.array(sll).reshape(len(sll),1)
    clf = IsolationForest(random_state=0).fit(X)
    predict_result=clf.predict(X)
    tl=[]
    for i in range(len(sll)):
        if predict_result[i] == 1:
            tl.append(sll[i])
    if len(tl) > 0:
        thresholds = pd.DataFrame({groupby_sdt : [name],'{}_threshold'.format(groupby_sdt):[max(tl)]})
        return thresholds

if __name__ == "__main__":
    cate_intervals = applyParallel(user_logs.groupby('user_id'), gen_intervals, 'cate')
    cate_thresholds = applyParallel(cate_intervals.groupby('cate'), iforest_remove_outliner, 'cate')
    pd.to_pickle(cate_thresholds.set_index('cate'), '../fpro_data/cate_intervals.pkl')
    brand_intervals = applyParallel(user_logs.groupby('user_id'), gen_intervals, 'store')
    brand_thresholds = applyParallel(brand_intervals.groupby('store'), iforest_remove_outliner, 'store')
    pd.to_pickle(brand_thresholds.set_index('store'), '../fpro_data/store_intervals.pkl')

brand_thresholds.set_index('store')

cate_thresholds.set_index('cate')
